Parse long instead of int for offset, as setXRef handles long offsets just fine, and the number is sometimes bigger than an int can handle.

+++ library/src/main/java/com/tom_roush/pdfbox/pdfparser/COSParser.java
@@ -1968,7 +1968,7 @@ public class COSParser extends BaseParser
                 {
                     try
                     {
-                        int currOffset = Integer.parseInt(splitString[0]);
+                        long currOffset = Long.parseLong(splitString[0]);
                         int currGenID = Integer.parseInt(splitString[1]);
                         COSObjectKey objKey = new COSObjectKey(currObjID, currGenID);
                         xrefTrailerResolver.setXRef(objKey, currOffset);
